Add models for connect server (See #80)
