+++ app/src/main/java/com/team3/fastcampus/record/Account/Modal/SignInData.java
@@ -1,4 +1,4 @@
-package com.team3.fastcampus.record.Account.Domain;
+package com.team3.fastcampus.record.Account.Modal;
 
 /**
  * Created by tokijh on 2017. 4. 5..
