+++ app/src/main/java/com/team3/fastcampus/record/Account/SigninActivity.java
@@ -29,8 +29,8 @@ import com.google.android.gms.common.SignInButton;
 import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.gson.JsonSyntaxException;
 import com.jakewharton.rxbinding2.widget.RxTextView;
-import com.team3.fastcampus.record.Account.Domain.SignInData;
-import com.team3.fastcampus.record.Account.Domain.SignUpData;
+import com.team3.fastcampus.record.Account.Modal.SignInData;
+import com.team3.fastcampus.record.Account.Modal.SignUpData;
 import com.team3.fastcampus.record.R;
 import com.team3.fastcampus.record.Util.Logger;
 import com.team3.fastcampus.record.Util.NetworkController;
