--- app/src/main/java/com/team3/fastcampus/record/InDiary/Domain/Image.java
@@ -1,10 +1,0 @@
-package com.team3.fastcampus.record.InDiary.Domain;
-
-/**
- * Created by tokijh on 2017. 4. 4..
- */
-
-public class Image {
-    public long id;
-    public byte[] imageData;
-}
