+++ app/src/main/java/com/team3/fastcampus/record/InDiary/Modal/Image.java
@@ -1,0 +1,13 @@
+package com.team3.fastcampus.record.InDiary.Modal;
+
+/**
+ * Created by tokijh on 2017. 4. 4..
+ */
+
+public class Image {
+    public long pk;
+    public long post;
+    public String photo;
+    public String gpsLatitude;
+    public String gpsLongitude;
+}
